+++ packages/adapters/src/adapters/Cornerstone3D/BaseAdapter3D.ts
@@ -3,6 +3,7 @@ import MeasurementReport, {
     type AdapterOptions,
     type MeasurementAdapter
 } from "./MeasurementReport";
+import { toScoords } from "../helpers";
 
 export type Point = {
     x: number;
@@ -142,55 +143,29 @@ export default class BaseAdapter3D {
 
     public static getTID300RepresentationArguments(
         tool,
-        worldToImageCoords
+        is3DMeasurement = false
     ): TID300Arguments {
-        const { data, metadata } = tool;
+        const { metadata } = tool;
         const { finding, findingSites } = tool;
         const { referencedImageId } = metadata;
-
-        if (!referencedImageId) {
-            return this.getTID300RepresentationArgumentsSCOORD3D(tool);
-        }
-
-        const {
-            handles: { points = [] }
-        } = data;
+        const scoordProps = {
+            is3DMeasurement,
+            referencedImageId
+        };
 
         // Using image coordinates for 2D points
-        const pointsImage = points.map(point => {
-            const pointImage = worldToImageCoords(referencedImageId, point);
-            return {
-                x: pointImage[0],
-                y: pointImage[1]
-            };
-        });
+        const pointsImage = toScoords(scoordProps, tool.data.handles.points);
 
         const tidArguments = {
             points: pointsImage,
             trackingIdentifierTextValue: this.trackingIdentifierTextValue,
             findingSites: findingSites || [],
-            finding
+            finding,
+            ReferencedFrameOfReferenceUID: is3DMeasurement
+                ? metadata.FrameOfReferenceUID
+                : null
         };
 
         return tidArguments;
-    }
-
-    static getTID300RepresentationArgumentsSCOORD3D(tool): TID300Arguments {
-        const { data, finding, findingSites } = tool;
-        const {
-            handles: { points = [] }
-        } = data;
-
-        // Using world coordinates for 3D points
-        const point = points[0];
-
-        const pointXYZ = { x: point[0], y: point[1], z: point[2] };
-
-        return {
-            points: [pointXYZ],
-            trackingIdentifierTextValue: this.trackingIdentifierTextValue,
-            findingSites: findingSites || [],
-            finding
-        };
     }
 }