+++ packages/adapters/src/adapters/Cornerstone3D/MeasurementReport.ts
@@ -1,13 +1,31 @@
-import { normalizers, data, utilities, derivations } from "dcmjs";
-import { cache } from "@cornerstonejs/core";
+import {
+    normalizers,
+    data,
+    utilities as dcmjsUtilities,
+    derivations
+} from "dcmjs";
+import {
+    cache,
+    utilities as csUtilities,
+    type Types as CSTypes
+} from "@cornerstonejs/core";
+import type { Types } from "@cornerstonejs/tools";
 
 import CORNERSTONE_3D_TAG from "./cornerstone3DTag";
-import { toArray, codeMeaningEquals, copyStudyTags } from "../helpers";
+import {
+    toArray,
+    codeMeaningEquals,
+    copyStudyTags,
+    scoordToWorld
+} from "../helpers";
 import Cornerstone3DCodingScheme from "./CodingScheme";
 import { copySeriesTags } from "../helpers/copySeriesTags";
+import { toPoint3 } from "../helpers/toPoint3";
 import { NO_IMAGE_ID } from "./constants";
 
-const { TID1500, addAccessors } = utilities;
+type Annotation = Types.Annotation;
+
+const { TID1500, addAccessors } = dcmjsUtilities;
 
 const { StructuredReport } = derivations;
 
@@ -24,28 +42,36 @@ const FINDING_SITE_OLD = { CodingSchemeDesignator: "SRT", CodeValue: "G-C0E3" };
 type SpatialCoordinatesState = {
     description?: string;
     sopInstanceUid?: string;
-    annotation: {
-        annotationUID: string;
-        metadata: {
-            toolName: string;
-            referencedImageId?: string;
-            FrameOfReferenceUID: string;
-            label: string;
-        };
-        data?: unknown;
-    };
+    annotation: Annotation;
     finding?: unknown;
     findingSites?: unknown;
 };
 
+type ScoordType = {
+    GraphicData: number[];
+};
+
 type SetupMeasurementData = {
     defaultState: SpatialCoordinatesState;
-    NUMGroup: Record<string, unknown>;
-    SCOORDGroup?: Record<string, unknown>;
+    state?: SpatialCoordinatesState;
+    is3DMeasurement?: boolean;
+    scoord?: ScoordType;
+    worldCoords?: CSTypes.Point3[];
+    scoordArgs?: {
+        referencedImageId: string;
+        is3DMeasurement: boolean;
+    };
+    NUMGroup: {
+        MeasuredValueSequence: {
+            NumericValue: number;
+        };
+    };
+    SCOORDGroup?: ScoordType;
     ReferencedSOPSequence?: Record<string, unknown>;
     ReferencedSOPInstanceUID?: string;
+    referencedImageId?: string;
     ReferencedFrameNumber?: string;
-    SCOORD3DGroup?: Record<string, unknown>;
+    SCOORD3DGroup?: ScoordType;
     FrameOfReferenceUID?: string;
 };
 
@@ -96,7 +122,6 @@ export interface MeasurementAdapter {
     getMeasurementData(
         measurementGroup,
         sopInstanceUIDToImageIdMap,
-        imageToWorldCoords,
         metadata,
         trackingIdentifier: string
     );
@@ -105,7 +130,7 @@ export interface MeasurementAdapter {
 
     getTID300RepresentationArguments(
         tool,
-        worldToImageCoords
+        is3DMeasurement
     ): Record<string, unknown>;
 }
 
@@ -128,11 +153,11 @@ export default class MeasurementReport {
         tool,
         ReferencedSOPSequence,
         toolClass,
-        worldToImageCoords
+        is3DMeasurement
     ) {
         const args = toolClass.getTID300RepresentationArguments(
             tool,
-            worldToImageCoords
+            is3DMeasurement
         );
         args.ReferencedSOPSequence = ReferencedSOPSequence;
 
@@ -159,7 +184,7 @@ export default class MeasurementReport {
         toolType,
         toolData,
         ReferencedSOPSequence,
-        worldToImageCoords
+        is3DMeasurement
     ) {
         const toolTypeData = toolData[toolType];
         const toolClass = this.measurementAdapterByToolType.get(toolType);
@@ -179,7 +204,7 @@ export default class MeasurementReport {
                 tool,
                 ReferencedSOPSequence,
                 toolClass,
-                worldToImageCoords
+                is3DMeasurement
             );
         });
 
@@ -248,7 +273,7 @@ export default class MeasurementReport {
         toolType,
         sopInstanceUIDToImageIdMap,
         metadata
-    }): SpatialCoordinatesData {
+    }) {
         const { ReferencedSOPSequence } = SCOORDGroup.ContentSequence;
         const { ReferencedSOPInstanceUID, ReferencedFrameNumber } =
             ReferencedSOPSequence;
@@ -260,22 +285,26 @@ export default class MeasurementReport {
             referencedImageId
         );
 
+        const annotationUID = DicomMetaDictionary.uid();
         return {
             SCOORDGroup,
             ReferencedSOPSequence,
             ReferencedSOPInstanceUID,
             ReferencedFrameNumber,
+            referencedImageId,
             state: {
                 description: undefined,
                 sopInstanceUid: ReferencedSOPInstanceUID,
                 annotation: {
-                    annotationUID: DicomMetaDictionary.uid(),
+                    data: {
+                        annotationUID
+                    },
+                    annotationUID,
                     metadata: {
                         toolName: toolType,
                         referencedImageId,
                         FrameOfReferenceUID:
-                            imagePlaneModule.frameOfReferenceUID,
-                        label: ""
+                            imagePlaneModule.frameOfReferenceUID
                     }
                 }
             }
@@ -286,22 +315,31 @@ export default class MeasurementReport {
         SCOORD3DGroup,
         toolType
     }): SpatialCoordinatesData {
-        return {
+        const annotationUID = DicomMetaDictionary.uid();
+        const toolData = {
             SCOORD3DGroup,
             FrameOfReferenceUID: SCOORD3DGroup.ReferencedFrameOfReferenceUID,
             state: {
                 description: undefined,
                 annotation: {
-                    annotationUID: DicomMetaDictionary.uid(),
+                    annotationUID,
+                    data: {
+                        annotationUID
+                    },
                     metadata: {
                         toolName: toolType,
                         FrameOfReferenceUID:
-                            SCOORD3DGroup.ReferencedFrameOfReferenceUID,
-                        label: ""
+                            SCOORD3DGroup.ReferencedFrameOfReferenceUID
                     }
                 }
             }
         };
+        csUtilities.updatePlaneRestriction(
+            toPoint3(SCOORD3DGroup.GraphicData),
+            toolData.state.annotation.metadata
+        );
+
+        return toolData;
     }
 
     public static getSpatialCoordinatesState({
@@ -346,7 +384,8 @@ export default class MeasurementReport {
             ReferencedSOPInstanceUID,
             ReferencedFrameNumber,
             SCOORD3DGroup,
-            FrameOfReferenceUID
+            FrameOfReferenceUID,
+            referencedImageId
         } = this.getSpatialCoordinatesState({
             NUMGroup,
             sopInstanceUIDToImageIdMap,
@@ -361,25 +400,27 @@ export default class MeasurementReport {
             return addAccessors(fsg.ConceptCodeSequence);
         });
 
-        const defaultState = {
-            ...state,
-            finding,
-            findingSites
-        };
+        state.finding = finding;
+        state.findingSites = findingSites;
 
-        if (defaultState.finding) {
-            defaultState.description = defaultState.finding.CodeMeaning;
+        if (finding) {
+            state.description = finding.CodeMeaning;
         }
 
-        defaultState.annotation.metadata.label =
-            MeasurementReport.getCornerstoneLabelFromDefaultState(defaultState);
+        state.annotation.data.label =
+            MeasurementReport.getCornerstoneLabelFromDefaultState(state);
 
         return {
-            defaultState,
+            // Deprecating the defaultState in favour of state, but there are lots
+            // of adapters still using defaultState
+            defaultState: state,
+            state,
             NUMGroup,
+            scoord: SCOORD3DGroup || SCOORDGroup,
             SCOORDGroup,
             ReferencedSOPSequence,
             ReferencedSOPInstanceUID,
+            referencedImageId,
             ReferencedFrameNumber,
             SCOORD3DGroup,
             FrameOfReferenceUID
@@ -406,7 +447,7 @@ export default class MeasurementReport {
             group => group.ValueType === "NUM"
         );
 
-        return this.processSpatialCoordinatesGroup({
+        const spatialGroup = this.processSpatialCoordinatesGroup({
             NUMGroup,
             sopInstanceUIDToImageIdMap,
             metadata,
@@ -414,6 +455,23 @@ export default class MeasurementReport {
             findingSiteGroups,
             toolType
         });
+
+        const { referencedImageId } = spatialGroup.state.annotation.metadata;
+        const is3DMeasurement = !!spatialGroup.SCOORD3DGroup;
+        const scoordArgs = {
+            referencedImageId,
+            is3DMeasurement
+        };
+        const scoord = spatialGroup.SCOORD3DGroup || spatialGroup.SCOORDGroup;
+        const worldCoords = scoordToWorld(scoordArgs, scoord);
+
+        return {
+            ...spatialGroup,
+            is3DMeasurement,
+            scoordArgs,
+            scoord,
+            worldCoords
+        };
     }
 
     static generateReferencedSOPSequence({
@@ -480,16 +538,14 @@ export default class MeasurementReport {
         const referenceToolData = toolData?.[toolTypes?.[0]]?.data?.[0];
         const volumeId = referenceToolData?.metadata?.volumeId;
         const volume = cache.getVolume(volumeId);
+        if (!volume) {
+            throw new Error(`No volume found for ${volumeId}`);
+        }
         const imageId = volume.imageIds[0];
         return imageId;
     }
 
-    static generateReport(
-        toolState,
-        metadataProvider,
-        worldToImageCoords,
-        options
-    ) {
+    static generateReport(toolState, metadataProvider, options) {
         // ToolState for array of imageIDs to a Report
         // Assume Cornerstone metadata provider has access to Study / Series / Sop Instance UID
         let allMeasurementGroups = [];
@@ -511,6 +567,7 @@ export default class MeasurementReport {
         Object.keys(toolState).forEach(imageId => {
             const toolData = toolState[imageId];
             const toolTypes = Object.keys(toolData);
+            const is3DMeasurement = imageId === NO_IMAGE_ID;
 
             const ReferencedSOPSequence = this.generateReferencedSOPSequence({
                 toolData,
@@ -521,7 +578,7 @@ export default class MeasurementReport {
                 derivationSourceDatasets
             });
 
-            if (imageId === NO_IMAGE_ID) {
+            if (is3DMeasurement) {
                 is3DSR = true;
             }
 
@@ -533,7 +590,7 @@ export default class MeasurementReport {
                     toolType,
                     toolData,
                     ReferencedSOPSequence,
-                    worldToImageCoords
+                    is3DMeasurement
                 );
                 if (group) {
                     measurementGroups.push(group);
@@ -564,6 +621,13 @@ export default class MeasurementReport {
         if (is3DSR) {
             report.dataset.SOPClassUID =
                 DicomMetaDictionary.sopClassUIDsByName.Comprehensive3DSR;
+            if (!report.dataset.SOPClassUID) {
+                throw new Error(
+                    `NO sop class defined for Comprehensive3DSR in ${JSON.stringify(
+                        DicomMetaDictionary.sopClassUIDsByName
+                    )}`
+                );
+            }
         }
 
         return report;
@@ -575,7 +639,6 @@ export default class MeasurementReport {
     static generateToolState(
         dataset,
         sopInstanceUIDToImageIdMap,
-        imageToWorldCoords,
         metadata,
         hooks
     ) {
@@ -644,7 +707,6 @@ export default class MeasurementReport {
                     const measurement = toolAdapter.getMeasurementData(
                         measurementGroup,
                         sopInstanceUIDToImageIdMap,
-                        imageToWorldCoords,
                         metadata,
                         trackingIdentifierValue
                     );