+++ packages/adapters/src/adapters/Cornerstone3D/UltrasoundDirectional.ts
@@ -1,8 +1,10 @@
 import { utilities } from "dcmjs";
+import { utilities as csUtilities } from "@cornerstonejs/core";
 import MeasurementReport from "./MeasurementReport";
 import BaseAdapter3D from "./BaseAdapter3D";
 
 const { Length: TID300Length } = utilities.TID300;
+const { worldToImageCoords } = csUtilities;
 
 class UltrasoundDirectional extends BaseAdapter3D {
     static {
@@ -39,6 +41,7 @@ class UltrasoundDirectional extends BaseAdapter3D {
         const state = defaultState;
 
         state.annotation.data = {
+            ...state.annotation.data,
             handles: {
                 points: [worldCoords[0], worldCoords[1]],
                 activeHandleIndex: 0,
@@ -46,14 +49,13 @@ class UltrasoundDirectional extends BaseAdapter3D {
                     hasMoved: false
                 }
             },
-            cachedStats: {},
             frameNumber: ReferencedFrameNumber
         };
 
         return state;
     }
 
-    static getTID300RepresentationArguments(tool, worldToImageCoords) {
+    static getTID300RepresentationArguments(tool, is3DMeasurement) {
         const { data, finding, findingSites, metadata } = tool;
         const { handles } = data;